Fix issues with alarm scheduling
