+++ apps/alarm/ChangeLog
@@ -1,1 +1,2 @@
 0.01: New App!
+0.02: Fix issues with alarm scheduling
