+++ apps/boot/ChangeLog
@@ -3,3 +3,4 @@
 0.04: Add alarm functionality
 0.05: Add Welcome screen on boot
 0.06: Disable GPS time log messages, add (default=1) setting to hide log messages
+0.07: Fix issues with alarm scheduling
